Move `HbarUnit` to its own file
    - Deprecate `Hbar.zero()`
